# v0.2/openapi.v2.json
{
  "components": {
    "schemas": {
...
      "ArtifactDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact": {
            "$ref": "#/components/schemas/ArtifactSummary"
          },
          "context": {
            "$ref": "#/components/schemas/ApiContext"
          },
          "direct_dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencySummary"
            },
            "title": "Direct Dependencies",
            "type": "array"
          },
          "direct_dependents": {
            "items": {
              "$ref": "#/components/schemas/DependencySummary"
            },
            "title": "Direct Dependents",
            "type": "array"
          },
          "lineage_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lineage Url"
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySummary"
          }
        },
        "required": [
          "context",
          "quality",
          "artifact",
          "direct_dependencies",
          "direct_dependents",
          "lineage_url"
        ],
        "title": "ArtifactDetailResponse",
        "type": "object"
      },
      "ArtifactListResponse": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/ApiContext"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtifactSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySummary"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "context",
          "quality",
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ArtifactListResponse",
        "type": "object"
      },
...
      "AssetCatalogResponse": {
        "additionalProperties": false,
        "properties": {
          "as_of_date": {
            "title": "As Of Date",
            "type": "string"
          },
          "context": {
            "$ref": "#/components/schemas/ApiContext"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssetCatalogItem"
            },
            "title": "Items",
            "type": "array"
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySummary"
          },
          "release_artifact_id": {
            "format": "uuid",
            "title": "Release Artifact Id",
            "type": "string"
          },
          "release_version_number": {
            "title": "Release Version Number",
            "type": "integer"
          },
          "universe_key": {
            "title": "Universe Key",
            "type": "string"
          }
        },
        "required": [
          "context",
          "quality",
          "release_artifact_id",
          "release_version_number",
          "as_of_date",
          "universe_key",
          "items"
        ],
        "title": "AssetCatalogResponse",
        "type": "object"
      },
...
      "DataOverviewResponse": {
        "additionalProperties": false,
        "properties": {
          "bundle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataBundleItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "context": {
            "$ref": "#/components/schemas/ApiContext"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetPublicationItem"
            },
            "title": "Datasets",
            "type": "array"
          },
          "eligibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EligibilitySnapshotItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySummary"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceSnapshotItem"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "context",
          "quality",
          "sources",
          "datasets",
          "bundle",
          "eligibility"
        ],
        "title": "DataOverviewResponse",
        "type": "object"
      },
...
      "DataRequirementResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "format": "uuid",
            "title": "Artifact Id",
            "type": "string"
          },
          "context": {
            "$ref": "#/components/schemas/ApiContext"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DataRequirementItem"
            },
            "title": "Items",
            "type": "array"
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySummary"
          },
          "requirement_set_key": {
            "title": "Requirement Set Key",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "context",
          "quality",
          "artifact_id",
          "requirement_set_key",
          "version_number",
          "items"
        ],
        "title": "DataRequirementResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LineageManifestResponse": {
        "additionalProperties": false,
        "properties": {
          "artifacts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "canonical_version": {
            "title": "Canonical Version",
            "type": "string"
          },
          "context": {
            "$ref": "#/components/schemas/ApiContext"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencySummary"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "manifest_hash": {
            "title": "Manifest Hash",
            "type": "string"
          },
          "quality": {
            "$ref": "#/components/schemas/QualitySummary"
          },
          "root_artifact": {
            "$ref": "#/components/schemas/ArtifactSummary"
          }
        },
        "required": [
          "context",
          "quality",
          "root_artifact",
          "manifest_hash",
          "canonical_version",
          "artifacts",
          "dependencies",
          "created_at"
        ],
        "title": "LineageManifestResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read-only v0.2 research identity, quality, and lineage interface.",
    "title": "Style Rotation Research API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v2/artifacts": {
      "get": {
        "operationId": "artifacts_api_v2_artifacts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "artifact_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifacts",
        "tags": [
          "lineage"
        ]
      }
    },
    "/api/v2/artifacts/{artifact_id}": {
      "get": {
        "operationId": "artifact_detail_api_v2_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifact Detail",
        "tags": [
          "lineage"
        ]
      }
    },
    "/api/v2/artifacts/{artifact_id}/lineage": {
      "get": {
        "operationId": "lineage_api_v2_artifacts__artifact_id__lineage_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lineage",
        "tags": [
          "lineage"
        ]
      }
    },
...
    "/api/v2/catalog/assets": {
      "get": {
        "operationId": "asset_catalog_api_v2_catalog_assets_get",
        "parameters": [
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset Catalog",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v2/catalog/data-requirements": {
      "get": {
        "operationId": "data_requirements_api_v2_catalog_data_requirements_get",
        "parameters": [
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRequirementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Data Requirements",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v2/data/overview": {
      "get": {
        "operationId": "data_overview_api_v2_data_overview_get",
        "parameters": [
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Data Overview",
        "tags": [
          "data"
        ]
      }
    },
    "/api/v2/factors/overview": {
      "get": {
        "operationId": "factor_overview_api_v2_factors_overview_get",
        "parameters": [
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactorOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Factor Overview",
        "tags": [
          "factor"
        ]
      }
    },
...
  }
}